Enhancement: Extract StringEndsWithTest
